Resolved last warning and removed abstract class.
